+++ Modules/Intent.Modules.Blazor/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 1.0.4
 
 - Improvement: Updated NuGet package versions.
+- Fixed: When creating proxy services from services not defined in a folder, error no longer occurs.
 
 ### Version 1.0.3